IONOS(ionos-mail): feat(controller): enhance PageController to identify IONOS-managed accounts
Added functionality to the PageController to determine if email accounts are managed by IONOS. This includes integrating the IonosMailService to fetch the user's IONOS email and updating the JSON response to indicate whether each account is IONOS-managed.

This improvement allows for better account management and user experience in the application.

Signed-off-by: Misha M.-Kupriyanov <kupriyanov@strato.de>